requirements.txt
prettytable==3.8.0
python-dateutil==2.8.2
validators==0.28.3
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6